Adding small changes

+++ Dog_App/data_prep.py
@@ -1,12 +1,7 @@
-import numpy as np
 import os
 import re
 import shutil
-import tqdm
-from sklearn.datasets import load_files
 from sklearn.model_selection import train_test_split
-from keras.preprocessing import image
-from keras.utils import np_utils
 from os.path import abspath
 
 
@@ -14,24 +9,28 @@ def _clean_dirnames(path, dl_folder):
     '''Clean the names of dog breed directories'''
     path = f"{path}/{dl_folder}"
     for file in os.listdir(path):
-        new_file = re.sub(" dog$", "", file)
+        new_file = re.sub(" dog$", "", file).replace(' ', '_')
         os.rename(f"{path}/{file}",
                   f"{path}/{new_file}")
 
 
 def _clean_files(path, dl_folder):
-    '''Remove image files that are .ash'''
+    '''Remove spaces and image files that are .ash'''
     path = f"{path}/{dl_folder}"
     for directory in os.listdir(path):
-        ash_files = list(filter(lambda x: re.search(".ash$", x),
-                                os.listdir(f"{path}/{directory}")))
-        for file in ash_files:
-            os.remove(f"{path}/{directory}/{file}")
+        for file in os.listdir(f"{path}/{directory}"):
+            if re.search(".ash$", file) is not None:
+                os.remove(f"{path}/{directory}/{file}")
+                continue
+            new_file = re.sub("^\d*\. ", '', file)
+            os.rename(f"{path}/{directory}/{file}",
+                      f"{path}/{directory}/{new_file}")
+
 
 
 def _create_partition_dirs(path):
     '''Create the train/valid/test folders'''
-    for directory in ["train", "validation", "test"]:
+    for directory in ["train", "valid", "test"]:
         os.mkdir(f"{path}/{directory}")
 
 
@@ -51,28 +50,31 @@ def _partition_data(path, dl_folder):
         pictures = os.listdir(abspath(f"{path}/{breed}"))
         train, test = train_test_split(pictures, train_size=0.8,
                                        random_state=23)
-        validation, test = train_test_split(test, train_size=0.5,
+        valid, test = train_test_split(test, train_size=0.5,
                                             random_state=23)
 
-        for directory in ["train", "validation", "test"]:
+        for directory in ["train", "valid", "test"]:
             _move_files(eval(directory), breed,
                         f"{path}/{breed}", f"{orig_path}/{directory}")
-
-    # Remove empty directory
-    shutil.rmtree(path)
 
 
 def setup_data(path, dl_folder):
     '''Public function to handle setting data up'''
+    if os.path.exists('dog_images'):
+        raise Exception('dog images folder already exists!')
+    else:
+        os.mkdir("dog_images")
+
     _clean_dirnames(path, dl_folder)
     _clean_files(path, dl_folder)
     _create_partition_dirs(path)
     _partition_data(path, dl_folder)
 
-    os.mkdir("dog_images")
-
-    for directory in ["train", "validation", "test"]:
+    for directory in ["train", "valid", "test"]:
         os.rename(directory, f"dog_images/{directory}")
+
+    # Remove empty directory
+    shutil.rmtree(f"{path}/{dl_folder}")
 
     print("Data prep complete")
